Implement 2D array input and analysis in C++
This code ask the user to first enter number of rows and number of columns. it then asks user to input numbers according to rows and coloumns . it then displays all the numbers enterd by users along with their indexes. also calculate the maximum and th minimum number, and calculates the sum of diagonal elements if the enterd 2d array is a square matrix.